pyproject.toml
[project]
name = "voicevault"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.9.0",
    "datasets[audio]>=4.0.0",
    "evaluate>=0.4.5",
    "fastapi[standard]>=0.116.1",
    "ffmpeg>=1.4",
    "jiwer>=4.0.0",
    "openai-whisper>=20250625",
    "pydantic>=2.11.7",
    "python-magic>=0.4.27",
    "python-multipart>=0.0.20",
    "torchcodec>=0.5",
    "transformers>=4.54.1",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.0",
    "speechrecognition>=3.14.3",
    "jupyter>=1.1.1",
    "ipywidgets>=8.1.7",
]
